Change the thread ID type from long to int32_t to aid in future porting.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler_1.0@294 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ Threads/vjBaseThread.cpp
@@ -3,7 +3,7 @@
 #include <Threads/vjBaseThread.h>
 #include <Threads/vjThreadManager.h>
 
-long vjBaseThread::mNextThreadId = 0;
+int32_t vjBaseThread::mNextThreadId = 0;
 
 	// ---- Ouput operator ---- //
 ostream& operator<<(ostream& out, vjBaseThread* threadPtr)
